requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
pycodestyle==2.9.1
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
zipp==3.8.1
